refactor: Typed `StyledSelect` with `SelectItem`
